Refactoring : one call to find_binary(...) to get the unique exe path
